chore(webapp): simplify default app setup
