--- webapp/javascript/redux/reducers/apps.ts
@@ -1,63 +1,0 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { AppNames } from '@models/appNames';
-import { fetchAppNames } from '@pyroscope/services/appNames';
-import type { RootState } from '../store';
-import { addNotification } from './notifications';
-
-interface NamesState {
-  // Since the value populated from the server
-  // There's no 'loading'
-  appNames:
-    | { type: 'loaded'; data: AppNames }
-    | { type: 'reloading'; data: AppNames }
-    | { type: 'failed'; data: AppNames };
-}
-
-// Define the initial state using that type
-const initialState: NamesState = {
-  appNames: { type: 'loaded', data: (window as any).initialState.appNames },
-};
-
-export const reloadAppNames = createAsyncThunk(
-  'names/reloadAppNames',
-  async (_, thunkAPI) => {
-    // TODO, retries?
-    const res = await fetchAppNames();
-
-    if (res.isOk) {
-      return Promise.resolve(res.value);
-    }
-
-    thunkAPI.dispatch(
-      addNotification({
-        type: 'danger',
-        title: 'Failed to load app names',
-        message: 'message' in res.error ? res.error.message : 'Unknown error',
-      })
-    );
-
-    return Promise.reject(res.error);
-  }
-);
-
-export const namesSlice = createSlice({
-  name: 'names',
-  initialState,
-  reducers: {},
-  extraReducers: (builder) => {
-    builder.addCase(reloadAppNames.fulfilled, (state, action) => {
-      state.appNames = { type: 'loaded', data: action.payload };
-    });
-    builder.addCase(reloadAppNames.pending, (state) => {
-      state.appNames = { type: 'reloading', data: state.appNames.data };
-    });
-    builder.addCase(reloadAppNames.rejected, (state) => {
-      state.appNames = { type: 'failed', data: state.appNames.data };
-    });
-  },
-});
-
-export const selectAppNamesState = (state: RootState) => state.apps.appNames;
-export const selectAppNames = (state: RootState) => state.apps.appNames.data;
-
-export default namesSlice.reducer;
