+++ webapp/javascript/redux/store.ts
@@ -16,7 +16,6 @@ import rootReducer from './reducers';
 import history from '../util/history';
 
 import viewsReducer from './reducers/views';
-import appsReducer from './reducers/apps';
 import settingsReducer from './reducers/settings';
 import userReducer from './reducers/user';
 import continuousReducer, {
@@ -26,7 +25,6 @@ import serviceDiscoveryReducer from './reducers/serviceDiscovery';
 import uiStore, { persistConfig as uiPersistConfig } from './reducers/ui';
 
 const reducer = combineReducers({
-  apps: appsReducer,
   root: rootReducer,
   views: viewsReducer,
   settings: settingsReducer,
